moved dirs and adjusted imports
