+++ app/quotes/__init__.py
@@ -1,7 +1,3 @@
 from app.quotes.quotes import quotes
-from app.quotes.relationships import quote_rel
-from app.customers.relationships import customer_rel
-from app.locations.relationships import location_rel
-from app.places.relationships import place_rel
-from app.products.relationships import product_rel
-# from app.vendors.relationships import vendor_rel
+from app.quotes.job_quotes.relationships import quote_rel
+from app.quotes.products.relationships import product_rel
