+++ app/quotes/job_quotes/models.py
