+++ app/quotes/products/relationships.py
@@ -1,8 +1,7 @@
 from fastapi import HTTPException, Depends
 from fastapi.routing import APIRouter
 from app import auth
-
-from app.quotes.models import RelatedQuoteResponse, QuoteRelationshipsResponse
+from app.quotes.job_quotes.models import RelatedQuoteResponse, QuoteRelationshipsResponse
 
 product_rel = APIRouter(prefix='/products', tags=['products'])
