+++ app/quotes/quotes.py
@@ -1,72 +1,73 @@
+from typing import Annotated
 from fastapi import HTTPException, Depends
 from fastapi.routing import APIRouter
 from app import auth
-from app.quotes.models import (
+from app.quotes.job_quotes.models import (
     QuoteResponse,
     NewQuoteResourceObject,
     ExistingQuote,
     QuoteQuery 
 )
-from app.products.models import ProductResponse, NewProductResourceObject, ExistingProduct
+from app.quotes.products.models import ProductResponse, NewProductResourceObject, ExistingProduct
 
 quotes = APIRouter(prefix='/quotes', tags=['quotes'])
 
+QuotesPerm = Annotated[auth.VerifiedToken, Depends(auth.quotes_perms_present)]
+
 @quotes.get('', response_model=QuoteResponse)
 async def quote_collection(
+        token: QuotesPerm,
         query: QuoteQuery=Depends(), # type: ignore
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
     ) -> QuoteResponse:
     raise HTTPException(status_code=501)
 
 @quotes.get('/{quote_id}')
 async def quote(
+        token: QuotesPerm,
         quote_id: int,
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
     ) -> QuoteResponse:
     raise HTTPException(status_code=501)
 
 @quotes.delete('/{quote_id}')
 async def delete_quote(
+        token: QuotesPerm,
         quote_id: int,
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
     ) -> None:
     raise HTTPException(status_code=501)
 
 @quotes.delete('/{quote_id}/products/{product_id}')
 async def delete_product(
+        token: QuotesPerm,
         quote_id: int,
         product_id: int,
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
     ) -> None:
     raise HTTPException(status_code=501)
 
-
 @quotes.post('')
 async def new_quote(
+        token: QuotesPerm,
         body: NewQuoteResourceObject,
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
     ) -> QuoteResponse:
     raise HTTPException(status_code=501)
 
 @quotes.post('/{quote_id}/products')
 async def add_product(
-        body: NewProductResourceObject,
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
+        token: QuotesPerm,
     ) -> ProductResponse:
     raise HTTPException(status_code=501)
 
 @quotes.patch('/{quote_id}')
 async def modify_quote(
+        token: QuotesPerm,
         quote_id: int,
         body: ExistingQuote,
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
     ) -> QuoteResponse:
     raise HTTPException(status_code=501)
 
 @quotes.patch('/{quote_id}/products')
 async def modify_product(
+        token: QuotesPerm,
         quote_id: int,
         body: ExistingProduct,
-        token: auth.VerifiedToken = Depends(auth.authenticate_auth0_token)
     ) -> ProductResponse:
     raise HTTPException(status_code=501)
